Prefer unique_ptr to auto_ptr
We still need the move semantic of auto_ptr, but with
unique_ptr we can make it explicit with std::move().

Suggested by Rein that says:
So std::unique_ptr makes the move semantics explicit, which is a good
thing because it documents the effects more clearly. IMO, there are no
real uses for std::auto_ptr anymore that std::unique_ptr cannot do at
least as well. IIRC, the std committee has marked auto_ptr as deprecated
(which means it can be removed in future versions, so code shouldn't rely
on it).

No functional change.