Re-enable "Active reparenting"
Once a slave has finished searching, instead of returning to idle
loop waiting to be picked up by a new master, it looks for an active
split point, register itself as a split point's slave and starts
searching on it. The idea is that average idle time is greatly
reduced in this case.

No functional change.