Retire Space evaluation Space evaluation seems like a good, bue actually it's rather counter productive. Intuitively, I am convinced that this space evaluation is a bad idea, for a couple of reasons 1/ it overlaps far too much with mobility, thereby violating the "orthogonality principle" 2/ it incentivises weakening moves like f4 (when OO) or c4 (when OOO) for instance. These are typical ill-advised pawn storm plans that more often backfire than they help. More generally, it incentivises pushing pawns for the sake of pushing pawns, which is rather dubious positional chess
Intuition aside, let's look at statistics!

8000 games in 10"+0.1", using SF's book limited to 8 moves (*)
1950-1641-4409 [51.9%] LOS=100.0%

So it's not even a close call, and this Space eval is definitely defective.

(*) tradeoff: if the bookdepth limit is too low, then 8000 games can't be played without too many repeated games. So one has to play less games, which means less precision. And of course more than 8 means that we go too deep into the middle game, leaving no chance to the space evaluation code to kick in (although I doubt it kicks in even with a low bookdepth). Also 8 moves seems to be the standard ot rating lists, and what SF is trying to achieve is maximum elo on rating lists after all.
signature=4800387