Generalize close position detection
Generalize code to detect and evaluate
any chain of at least 3 pawns and not
only the central one.

The chain is scored according to its
topology properties and total score
is used to alter eval scale factor.

Current patch just introduces infrastructure
without enabling the scale factor, so
it is a non functional patch with master.

bench: 7905850