Rename UCI 'divide' command to 'perft'
The main purpose of perft is to help debugging. But without the
breakdown in sum of perft(N-1), it is a completely useless debugging
tool. So perft now displays the breakdown, and divide is therefore
removed.

To allow speed tests output has been redirected to std::cout so that a

./stockfish perft 6 > /dev/null

will not be impacted by console printing overhead.

No functional change.