Retire "Write Search Log" UCI option
It is an unused feature and a bit redundant with the
GUI output that shows already the same info.

As a side benefit, apart fom big code reduction,
we don't have anymore a different bench number with
search log on or off (due to pretty_pv).

While there sync Makefile and 'contempt' UCI option
with upstream.

No functional change.