Add support for UCI extension: nps NODE_RATE
When running more games in parallel, or simply when running a game with a
background process, due to how OS scheduling works, there is no guarantee
that the CPU resources allocated evenly between the two players.

This introduces noise in the result that leads to unreliable result and in
the worst cases can even invalidate the result.

For instance in SF test framework we avoid running from clouds virtual
machines because are a known source of very unstable CPU speed.

To overcome this limit, a simple extension to UCI protocol is proposed.

The idea is to use searched nodes instead of time, and to convert time
to/from nosed searched in as few places as possible.

Namely, we propose to add to the 'go' command definition in the UCI
protocol, the following line:

* nps NODE_RATE
tells the engine to use nodes searched instead of wall time to account
for elapsed time. Nodes searched will be converted to elapsed seconds
through dividing by the given NODE_RATE.

This patch implements the above functionality, engine side.

WARNING: Currently only single thread search is supported.

No functional change.