Fix off-by-one DTZ bug
For a given position DTZ tables store results for only one side to
move, so in case we have the wrong color, we perform a 1-ply search
and probe DTZ with the opposite color, then we return the minimum DTZ
compensated by one ply. But for zeroing moves, zeroing_move_dtz() is
_already_ one ply compensated because returns the DTZ of the previous
move before the zeroing move, so we should not convert back from 1-ply
search in this case.

This is a case in which the real bug is the misnaming of a function.

Indeed zeroing_move_dtz() should be renamed before_zeroing_move_dtz()
that is what it really returns. After the rename the bug becomes
obvious and is easily fixed.

Bug triggers with 6-men tables in this position:

8/8/7p/1K3N2/8/2b5/2b4k/8 w - - 5 7

Where DTZ returns -102 instead of the correct -101