Fix hash keys
The benches were different for standard chess when compiling with variants,
which was caused by changes to the PRNG state and XORing of three-check keys.

Furthermore, the hash keys for three-check positions could be different
for same positions and the same for different positions.

Both should be fixed by this commit.