Retire piece lists
On modern CPU lsb() instruction is executed in hardware and
is very fast, this alow us to loop across pieces using a
bitboard and removing all teh piece list logic.

Note that in case of 32 bit CPU this patch introduces a slowdown,
but nowdays this is no more main target and wth new dev cycle it
is right time to get rid of this obsoleted code.

Original patch is from Lucas:

http://tests.stockfishchess.org/tests/view/567d313a0ebc592d552a4491

And it already passed as a simplification:

LLR: 2.96 (-2.94,2.94) [-3.00,1.00]
Total: 123190 W: 16586 L: 16603 D: 90001

Bench: 4986710