Make bindThisThread() thread-safe
We coudl have a teoretical race if multiple
threads call bindThisThread() while we are
initializing the static variable. To avoid any
issues, the simplest way is to avoid static
variables and use locals instead. The drawback
is that we are going to re-compute process info
once per thread. This is not an issue given
that this is startup code and we are talking of
a couple of millisecs of delay anyhow.