Find a longer PV within a certain range if the best PV is too short.
If the best chosen PV is shorter than a given number of plies, look for
a better PV to use which is within a certain score and depth range.
Ensure that the longer PV uses all of the same moves from the chosen PV.

Send it to the GUI after the chosen PV is sent.

This likely needs a bit of tunining and careful consideration, but it
should find a reasonable PV with perhaps a bit of loss of strength.

It probably needs to be heavily tested on the games that show the small
PVs to ensure that the chosen PV is worth showing.