Clarify search criteria, prefer higher eval once our PV is long enough
In addition to making the algorithm more readable (IMO), this fixes a
couple of issues raised during the review.

1. any PV that matches the bestThread PV is a candidate now. Previously
I was mistakenly requiring each subsequent thread to match the previous
longestPVThread, which was dependent on the order the threads showed
up. This is much better.

2. So long as our PV is smaller than the minPlies, we will take any
thread that provides a longer candidate PV and a score/depth that is
"close enough" to the bestThread.  Once our PV has reached minPlies, we
will take any thread that improves either score or depth, so long as it
also is minPlies long.

I think this is a nice middle ground, although the values for minPlies,
maxScoreDiff and maxDepthDiff should be tuned. I just chose 10
centipawns for score and 2 for depth based on some assumptiong I
personally made, but I wouldn't argue for them at all.  minPlies was
set to 6 based on feedback from Thibault, as ultimately we are looking
to satifisy Users perception in the UI, and so I wanted to ensure I was
fixing that.

Also note that this is not guaranteed to provide a longer PV. If all
threads have short PVs (not sure if that will ever happen) then we may
still get a shorter PV.  I hope to be able to work on the Transposition
Table PV extraction over the holidays.