Allow variant as optional argument to bench command
E.g. "bench atomic". If the word after bench starts with a digit,
the current UCI_Variant option value is used and the word
is instead passed to the next option to preserve compatibility
to bench commands of official stockfish.

Furthermore, add an array of default depths to be able to account
for different run times of benchmarks in variants.