Rewrite move encoding (resolves #103)
Rewrite move encoding of piece drops and king promotions
in order to fit into 16 bits, so that TT works properly.

The bits for the move type are set as if it was a normal move,
and the bits for the promotion piece type are used to set
flags for piece drops and king promotions.

The dropped piece is encoded in the bits for the origin square.

crazyhouse:
STC
LLR: 2.96 (-2.94,2.94) [0.00,10.00]
Total: 618 W: 348 L: 241 D: 29

LTC:
LLR: 2.95 (-2.94,2.94) [0.00,10.00]
Total: 458 W: 264 L: 162 D: 32

giveaway:
STC
LLR: 2.96 (-2.94,2.94) [-10.00,5.00]
Total: 4202 W: 1103 L: 1080 D: 2019

chess:
STC
LLR: -3.00 (-2.94,2.94) [-10.00,5.00]
Total: 11952 W: 2145 L: 2239 D: 7568

Speed regression in standard chess to be fixed by follow-up commit.