Compensate elapsed time ans set slowdown at 1%
Increase the elapsed time used to check the finish of
available time of a quanity corresponding to the required
slowdown.

This will make the search to finish earlier so that when
at the end of the search we wait for the slowdown, the total
time before returning is not affected and we don't create the
artifact of always exit the search later than expected.

Suggested by lp-- here
https://github.com/mcostalba/Stockfish/commit/2c987c331