Another attempt at slowdown code
This is a simpler solution that should work
withouth corrections but it is still based
on system clock, so it should be verified on
different OS/compilers.

It is based on slowing down after each Iterative
Deepening search, i.e. after each seacrh at increasing
depth. It is designe to work only with single thread.

Current version slowdowns of 4% compared to master.

Verified with:

./stockfish go wtime 80000 > /dev/null

No functional change.