Fix from_to for piece drops in crazyhouse (fixes #432)
Since the bits of the destination square are used to encode the piece type
of the dropped piece, we have to exclude these bits when determining
from_to for piece drops in order to avoid exceeding the limits, i.e.,
causing out of bound array accesses.

The failed assertion occured when the index for the mainHistory exceeded
the limits and actually changed a value in the adjacent captureHistory,
which has lower maximum values.