Fix generation of pawn promotions for losers chess
Since quiet moves are filtered out in move generation of captures
and vice versa, queen promotions are suppressed in move generation,
because they are handled as captures even if they are quiet moves.

Example:
8/p5P1/2p5/8/6k1/8/6K1/8 b - - 0 1
Position is mate in 2 if pawn is promoted to queen, but search only finds
mate in 3 by promotion to rook since search is blind to queen promotions.