z3d
the following are true:

A value of type IntPairTransform may be invoked with a single argument of type IntPair.
The result of invoking a function of type IntPairTransform may not be used where a value of type IntPair is expected.
A value of type IntPairTransform2 may be used when an IntPairTransform is expected, but not vice versa.
A value of type IntPairTransform may be used when an IntPairTransform3 is expected, but not vice versa.
Functors
A functor in Q# is a factory that defines a new operation from another operation. Functors have access to the implementation of the base operation when defining the implementation of the new operation. Thus, functors can perform more complex functions than traditional higher-level functions.

A functor is used by applying it to an operation, returning a new operation. For example, the operation that results from applying the Adjoint functor to the Y operation is written as (Adjoint Y). The new operation may then be invoked like any other operation. Thus, (Adjoint Y)(q1) applies the adjoint functor to the Y operation to generate a new operation, and applies that new operation to q1.

Similarly, (Controlled X)(controls, target)

The two standard functors in Q# are Adjoint and Controlled.

Adjoint
In quantum computing, the adjoint of an operation is the complex conjugate transpose of the operation. For operations that implement a unitary operator, the adjoint is the inverse of the operation. For a simple operation that just invokes a sequence of other unitary operations on a set of qubits, the adjoint may be computed by applying the adjoints of the sub-operations on the same qubits, in the reverse sequence.

Given an operation expression, a new operation expression may be formed using the Adjoint functor, with the base operation expression enclosed in parentheses, ( and ). The new operation has the same signature and type as the base operation. In particular, the new operation also allows Adjoint, and will allow Controlled if and only if the base operation did.

For instance, (Adjoint QFT) designates the adjoint of the QFT operation.

Controlled
The controlled version of an operation is a new operation that effectively applies the base operation only if all of the control qubits are in a specified state. If the control qubits are in superposition, then the base operation is applied coherently to the appropriate part of the superposition. Thus, controlled operations are often used to generate entanglement.

In Q#, controlled versions always take an array of control qubits, and the specified state is always for all of the control qubits to be in the computational (PauliZ) One state, 
|
1
?
. Controlling based on other states may be achieved by applying the appropriate Clifford operations to the control qubits before the controlled operation, and then applying the inverses of the Cliffords after the controlled operation. For example, applying an X operation to a control qubit before and after a controlled operation will cause the operation to control on the Zero state (
|
0
?
) for that qubit; applying an H operation will control on the PauliX Zero state 
|
+         
?
:=
(
|
0
?
+
|
1
?
)
   /
v
2